scripts/research_assistant/requirements.txt
# Research Assistant Dependencies
anthropic>=0.40.0
feedparser>=6.0.0
pyyaml>=6.0
